Supplier Zone (get by source ID) - remove AddTestCallInput, Output Class for Scheduler task - ZoneID long - remove InitiateTestInformation, TestProgress classes from CliTester.Entitites.CLITesterConnectorBase - Remove RetestCall fct - Add Grid to TestCallDetailsTemplate
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C30600

+++ QualityMeasurement/Code/QM.CLITester.Data/ITestCallDataManager.cs
@@ -9,7 +9,7 @@ namespace QM.CLITester.Data
 {
     public interface ITestCallDataManager : IDataManager
     {
-        bool Insert(int supplierId, int countryId, int zoneId, int callTestStatus, int callTestResult,
+        bool Insert(int supplierId, int countryId, long zoneId, int callTestStatus, int callTestResult,
             int initiationRetryCount, int getProgressRetryCount, int userId, int profileId, long? batchNumber, int? scheduleId);
         List<TestCall> GetUpdated(ref byte[] maxTimeStamp, int nbOfRows, int userId);
         List<TestCall> GetBeforeId(GetBeforeIdInput input);
