+++ QualityMeasurement/Code/QM.CLITester.Entities/TestCall.cs
@@ -12,7 +12,7 @@ namespace QM.CLITester.Entities
         public long ID { get; set; }
         public int SupplierID { get; set; }
         public int CountryID { get; set; }
-        public int ZoneID { get; set; }
+        public long ZoneID { get; set; }
         public int UserID { get; set; }
         public int ProfileID { get; set; }
         public DateTime CreationDate { get; set; }
